+++ pecrs/space.py
@@ -2,85 +2,77 @@
 from pecrs.spatialhash import SpatialHash
 from pecrs.collider import Collider
 
-
 class Space:
    """
    :param size: Size of the collision areas
    :type size: int
 
-   Space handles the interactions between the SpatialHash and the Collider.
+   Space handles the movement and collision detection between shapes.
    You can change the values here to work with different spatial partioning system and collision systems.
    """
 
-   def __init__(self, size):
+   def __init__(self, size=256):
       self.grid = SpatialHash(size) #: SpatialHash, spatial partitioning system. Think about combining these into one class
       self.collider = Collider() #: Collider, collision detection system
 
-   def add(self, body, area=None):
+   def add(self, shape, update_area=True):
       """
-      :param body: Body to add to the space
-      :param area: Area of the space to add the body to. Calculate from body.position if None
-      :type body: AbsBody
-      :type area: tuple(int, int) or None
+      :param shape: Shape to add to the space
+      :type shape: Shape
       
-      Adds a body to the space.
+      Adds a shape to the space.
       Bodies are added to their collision area as well as the 8 nearest neighboring cells to handle overlap and fast moving obejcts.
       """
-      if area:
-         x_area = area[0]
-         y_area = area[1]
-      else:
-         body.area = self.grid.scale(body.shape.position[0], body.shape.position[1])
-         x_area = body.area[0]
-         y_area = body.area[1]
+      if update_area:
+         shape.area = self.grid.scale(shape.position[0], shape.position[1])
 
-      self.grid.add(body, (x_area-1, y_area-1))
-      self.grid.add(body, (x_area, y_area-1))
-      self.grid.add(body, (x_area+1, y_area-1))
-      self.grid.add(body, (x_area-1, y_area))
-      self.grid.add(body, (x_area, y_area))
-      self.grid.add(body, (x_area+1, y_area))
-      self.grid.add(body, (x_area-1, y_area+1))
-      self.grid.add(body, (x_area, y_area+1))
-      self.grid.add(body, (x_area+1, y_area+1))
+      x_area = shape.area[0]
+      y_area = shape.area[1]
 
-   def delete(self, body, area=None):
+      self.grid.add(shape, (x_area-1, y_area-1))
+      self.grid.add(shape, (x_area, y_area-1))
+      self.grid.add(shape, (x_area+1, y_area-1))
+      self.grid.add(shape, (x_area-1, y_area))
+      self.grid.add(shape, (x_area, y_area))
+      self.grid.add(shape, (x_area+1, y_area))
+      self.grid.add(shape, (x_area-1, y_area+1))
+      self.grid.add(shape, (x_area, y_area+1))
+      self.grid.add(shape, (x_area+1, y_area+1))
+
+   def delete(self, shape):
       """
-      :param body: Body to removed from the space
-      :param area: Area of the space to remove the body from. Calculate from body.position if None
-      :type body: AbsBody
-      :type area: tuple(int, int) or None
+      :param shape: Shape to removed from the space
+      :type shape: Shape
 
-      Removes a body from the space.
+      Removes a shape from the space.
       """
-      if area:
-         x_area = area[0]
-         y_area = area[1]
-      else:
-         x_area = body.area[0]
-         y_area = body.area[1]
+      x_area = shape.area[0]
+      y_area = shape.area[1]
 
-      self.grid.delete(body, (x_area-1, y_area-1))
-      self.grid.delete(body, (x_area, y_area-1))
-      self.grid.delete(body, (x_area+1, y_area-1))
-      self.grid.delete(body, (x_area-1, y_area))
-      self.grid.delete(body, (x_area, y_area))
-      self.grid.delete(body, (x_area+1, y_area))
-      self.grid.delete(body, (x_area-1, y_area+1))
-      self.grid.delete(body, (x_area, y_area+1))
-      self.grid.delete(body, (x_area+1, y_area+1))
+      self.grid.delete(shape, (x_area-1, y_area-1))
+      self.grid.delete(shape, (x_area, y_area-1))
+      self.grid.delete(shape, (x_area+1, y_area-1))
+      self.grid.delete(shape, (x_area-1, y_area))
+      self.grid.delete(shape, (x_area, y_area))
+      self.grid.delete(shape, (x_area+1, y_area))
+      self.grid.delete(shape, (x_area-1, y_area+1))
+      self.grid.delete(shape, (x_area, y_area+1))
+      self.grid.delete(shape, (x_area+1, y_area+1))
 
-   def has(self, body):
+   def has(self, shape):
       """
-      :param body: Body to check for
-      :type body: AbsBody
-      :return: True if the space contains the body, False if not
+      :param shape: Shape to check for
+      :type shape: Shape
+      :return: True if the space contains the shape, False if not
       :rtype: bool
 
-      Checks if a space has a body in that body's area
-      Note that the body's must be set correctly or errors can occur
+      Checks if a space has a shape in that shape's area
+      Note that the shape's must be set correctly or errors can occur
       """
-      return self.grid.has(body, body.area)
+      try:
+         return self.grid.has(shape, shape.area)
+      except AttributeError:
+         return False
 
    def collisions_at(self, x, y, width=1, height=1):
       """
@@ -92,32 +84,31 @@ class Space:
       :type y: Int
       :type width: Int
       :type height: Int
-      :return: List of bodies colliding at the position
-      :rtype: list(AbsBody)
+      :return: List of shapes colliding at the position
+      :rtype: list(Shapes)
 
-      Gets a list of all bodies colliding at x, y
+      Gets a list of all shapes colliding at x, y
       """
       area = self.grid.scale(x, y)
       collisions = []
-      bucket = self.grid.get(area)
-      for body in bucket:
-         if self.collider.rect_rect(width, height, x, y, body.shape.width, body.shape.height, body.shape.position[0], body.shape.position[1]):
-            collisions.append(body)
+      for shape in self.grid.get(area):
+         if self.collider.rect_rect(width, height, x, y, shape.width, shape.height, shape.position[0], shape.position[1]):
+            collisions.append(shape)
       return collisions
 
-   def collisions_with(self, body):
+   def collisions_with(self, shape):
       """
-      :param body: Body to find collisions with
-      :type body: AbsBody
-      :return: List of bodies colliding at the position
-      :rtype: list(AbsBody)
+      :param shape: Shape to find collisions with
+      :type shape: Shape
+      :return: List of shapes colliding at the position
+      :rtype: list(Shape)
 
-      Get a list of all bodies colliding with body
+      Get a list of all shapes colliding with shape
       """
       collisions = []
-      for other in self.grid.get(body.area):
-         if body.id != other.id:
-            if self.collider.check_bodies(body, other):
+      for other in self.grid.get(shape.area):
+         if shape != other:
+            if self.collider.check_rects(shape, other):
                collisions.append(other)
       return collisions
       
@@ -137,51 +128,98 @@ class Space:
       Check to see if there is a collision at x, y in space.
       """
       area = self.grid.scale(x, y)
-      for body in self.grid.get(area):
-         if self.collider.rect_rect(width, height, x, y, body.width, body.height, body.position[0], body.position[1]):
+      for shape in self.grid.get(area):
+         if self.collider.rect_rect(width, height, x, y, shape.width, shape.height, shape.position[0], shape.position[1]):
             return True
       return False
       
-   def check_body(self, body):
+   def check(self, shape):
       """
-      :param body: Body to check for collisions
-      :type body: AbsBody
-      :return: True if the body is colliding with anything in space, False if not
+      :param shape: Shape to check for collisions
+      :type shape: Shape
+      :return: True if the shape is colliding with anything in space, False if not
       :rtype: bool
 
-      Check to see if something is colliding with body in the space.
+      Check to see if something is colliding with shape in the space.
       """
-      for other in self.grid.get(body.area):
-         if body.id != other.id:
-            if self.collider.check_bodies(body, other):
+      for other in self.grid.get(shape.area):
+         if shape != other:
+            if self.collider.check_rects(shape, other):
                return True
       return False
 
-   def check_bodies(self, body, other):
-      return self.collider.check_bodies(body, other)
-
-   def move(self, body, dir):
+   def check_two(self, shape, other):
       """
-      :param body: Body to add to the space
-      :param dir: Direction to move the body in
-      :type body: AbsBody
-      :type dir: Vector
+      :param shape: First Shape to check for collision
+      :param other: Second Shape to check for collision
+      :type shape: Shape
+      :type other: Shape
+      :return: True if the two shapes are colliding, False if not
+      :rtype: Bool
 
-      Moves a body. This will be more efficent than place
-      TODO Not implemented yet. Use place(body, area) instead
+      Check to see if two shapes are colliding
       """
-      return
+      return self.collider.check_rects(shape, other)
 
-   def place(self, body, area):
+   def update_area(self, shape):
       """
-      :param body: Body to be moved from the space
-      :param area: Area of the space to move the body to
-      :type body: AbsBody
+      :param shape: Body to be moved from the space
+      :param area: Area of the space to move the shape to
+      :type shape: AbsBody
       :type area: tuple(int, int)
 
-      Moves a body from one collision area to another
+      Moves a shape from one collision area to another
       """
-      self.delete(body, area)
-      body.area = area
-      self.add(body, area)
+      area = self.grid.scale(shape.position[0], shape.position[1])
+      if shape.area != area:
+         self.delete(shape)
+         shape.area = area
+         self.add(shape, False)
+      
+   def move(self, shape, x, y, delta):
+      """
+      :param shape: Shape to move
+      :param x: Direction to move in on horizontal plane
+      :param y: Direction to move in on the vertical plane
+      :param delta: Amount of time to move for
+      :type shape: Shape
+      :type x: int
+      :type y: iny
+      :type delta: float
+      :return: Distance moved
+      :rtype: tuple(int, int)
+      
+      Advances a shape in the direction x, y for delta units of time
+      """
+      xstep = int(x * delta)
+      ystep = int(y * delta)
+      shape.position = (shape.position[0]+xstep, shape.position[1]+ystep)
+      self.update_area(shape)
+      return (xstep, ystep)
 
+   def push(self, shape, x, y):
+      """
+      :param shape: Shape to push
+      :param x: Direction on the horizontal plane to push the shape
+      :param y: Direction on the vertical plane to push the shape
+      :type shape: Shape
+      :type x: int
+      :type y: int
+
+      Pushes a shape in the direction of x, y in space
+      """
+      shape.position = (shape.position[0]+x, shape.position[1]+y)
+      self.update_area(shape)
+
+   def place(self, shape, x, y):
+      """
+      :param x: The position in space to place the shape on the horizontal plane
+      :param y: The position in space to place the shape on the vertical plane
+      :type x: int
+      :type y: int
+
+      Directly places a shape at x, y in space
+      """
+      shape.position = (x, y)
+      self.update_area(shape)
+